From Chris:
There was a bit of a glitch in that title grabber extension I wrote to
the URL plugin -- namely, that it would just mindlessly download
whatever link anybody posted to the channel. Whenever anybody posted a
bigass quicktime movie or something, the bot would explode. So, I had
to get down and dirty with Net::HTTP to check for mimetypes and
redirects and stuff.

There is a nice new feature though:

22:37 <@epitron>
http://killswitchengage.com/shared/downloads/KillswitchEngage/KsE-RoseOfSharyn_v
ideo.mov
22:37 <@pookie> [Link Info] type: video/quicktime, size: 18,244,208 bytes