rss plugin: prevent deadlock between save and watcher threads.
Since the timer fix in [1035], rescheduling ticks the timer, which could cause the timed saving to trigger while a feed mutex was being held. Fix by moving the feed watcher rescheduling out of the mutex synchronization.

Also move other calculations out to hold the mutex for as little as necessary.