gettext: support anonymous modules (and thus plugins) with newer gettext
While older ruby-gettext raised when trying to bind for anonymous
modules (such as the ones that wrap rbot plugins), newer versions just
provide an empty array. This makes the previous fix (which used rescue)
ineffective.

Solve by rescuing with [] (so that older gettext behaves like the new
one) and then adding Object if the list is indeed empty.