basics: option to join channel after identification is confirmed
Sometimes it is necessary to wait for identification to be confirmed
before certain channels may be joined. In this case the option
irc.join_after_identify can be set to true, and the bot will wait for
nickserv to confirm the identification before joining any channels.

This solution is actually a rather ugly hack, but I can't think of a
better way to approach the problem without rewriting the whole
framework.