rss plugin: parseRss returns the number of found items
Change the parseRss routine to return the number of found items, or nil
in case of error. This helps clearly differentiate between empty feeds
(which are still legit) and b0rked feeds.

This change in logic does not alter the fact that a feed update with no
items will not wipe existing old feed items.