geoip: make the GeoIP module a standalone
The geoip plugin defines and makes use of a GeoIP module to handle the
actual geoip data retrieval. The module used to be defined in the same
namespace as the plugin, and therefore in a barely accessible anonymous
module.

Move the GeoIP module to the outermost namespace to allow it to be
accessible from scripts and plugins too.