gettext: support version 2
GetText version 2 has some significant difference from earlier version.

* different syntax to set the default locale path
* different syntax to set non-cached mode
* different way to handle bound targets

Most of the changes are relative to significant functionality split
between the GetText submodules (LocalePath, TextDomain etc), so most of
the changes are just a matter of moving the defines where appropriate.
The bound_targets patch needed to cope with anonymous modules is not
needed with gettext >= 2.0.0