Better handling of MessageMapper failures
The plugin fallback method is now passed the list of failures.

Failures themselves are encapsulated in their own data type, making it
easier to handle failures that should inform the user with something
more detailed than the classic 'usage' pattern.

It's still up to the fallback method to make use (e.g. echo) the
relevant messages.