Revert "utils: try nokogiri/hpricot if hpricot is not available"
This reverts commit 4ea25eba02d85b81dcee4302375d7d3e08cd52ac.

Not only the patch was missing an essential part, but Nokogiri is not
compatible enough with Hpricot anyway. Incompatibilities include
doc/"style|script" raising an "Undefined namespace prefix" error and the
non-existence of Hpricot::Elements.

If we want to support Nokogiri we'll need to write code for it on
purpose or redesign the Hpricot code so that it only uses the subset
supported by Nokogiri.