remind: fix presence of 'in', 'every' or 'at' in message
if 'in' was present in the message but not part of the time
specification (e.g. because the specification was with 'every' or with
'at'), the parser would fail at the first attempt (look for time after
'in') and fail unnecessarily.

Fix by trying all possibilities and only giving an error if none
succeeds.