weather: URI-encode station
The NWS code assumed that the station could be used as-is in the URL,
which is actually true for the station codes accepted by NWS, but
incorrect in general. People with NWS as the default weather service
would get strange errors about bad URIs if trying something like
"weather birmingham, uk".

Solve by URI-encoding the station name: even if the resulting URL will
still give a 404, at least it doesn't give out odd errors.