Mark threaded mapped messages as replied
This prevents them from being delegated to unreplied() before the processing
thread has the chance to mark it replied. The solution would fail for
threaded messages whose processing would not reply, but I'm not aware of
such a method currently exists.

A future design of the message delegation and mapping could thread the
actual delegation instead, and this problem would be solved.